# docs/API/openapi.yaml
openapi: 3.1.0
info:
  title: Genesis 360 Empresarial API
  version: 1.2.1-rc.1
  description: "Contratos HTTP da baseline V1. Autorização de tenant é validada no servidor; IDs enviados pelo cliente\
    \ nunca substituem o tenant ativo autenticado.\n V1.2.1: mutações por cookie exigem Origin igual ao host servido;\
    \ papéis autorizados por operação; corpo JSON limitado a 64 KiB. Material sensível é somente declarado pela\
    \ entrada pública."
servers:
- url: /api
...
paths:
...
  /tenant/active:
    post:
      tags:
      - Tenant
      operationId: setActiveTenant
      summary: Seleciona tenant ativo entre memberships autorizadas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - tenantId
              properties:
                tenantId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Tenant ativo selecionado
          content:
            application/json:
              schema:
                type: object
                required:
                - tenantId
                properties:
                  tenantId:
                    type: string
                    format: uuid
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Origem, tenant ou papel não autorizados; nunca expõe detalhes internos.
        '409':
          description: Conflito de estado/revisão; reler antes de tentar gravação.
        '413':
          description: Corpo acima de 64 KiB.
        '500':
          description: Falha interna sem vazamento de mensagens do provedor.
  /tenant/members:
    get:
      tags: [Tenant]
      operationId: listTenantMembers
      summary: Lista membros para owner/admin do tenant ativo
      responses:
        '200': { description: Membros do tenant }
        '403': { description: Acesso negado }
    post:
      tags: [Tenant]
      operationId: inviteTenantMember
      summary: Convida membro com papel controlado
      responses:
        '201': { description: Convite criado }
        '403': { description: Acesso negado }
        '409': { description: Conflito de ciclo de vida }
    patch:
      tags: [Tenant]
      operationId: updateTenantMemberRole
      summary: Altera papel preservando ao menos um owner
      responses:
        '200': { description: Papel alterado }
        '403': { description: Acesso negado }
        '409': { description: Último owner protegido }
    delete:
      tags: [Tenant]
      operationId: removeTenantMember
      summary: Revoga membership do tenant ativo
      responses:
        '200': { description: Acesso revogado }
        '403': { description: Acesso negado }
        '409': { description: Último owner protegido }
...
  /passport/facts:
    get:
      tags:
      - Passport
      operationId: listBusinessFacts
      summary: Lista fatos atuais do Business Passport
      parameters:
      - name: companyId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Fatos atuais
    post:
      tags:
      - Passport
      operationId: recordBusinessFact
      summary: Versiona um Business Fact por trusted server RPC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessFactInput'
      responses:
        '201':
          description: Nova versão criada
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: Origem, tenant ou papel não autorizados; nunca expõe detalhes internos.
        '409':
          description: Conflito de estado/revisão; reler antes de tentar gravação.
        '413':
          description: Corpo acima de 64 KiB.
        '500':
          description: Falha interna sem vazamento de mensagens do provedor.
...
  /missions/{id}/transition:
    post:
      tags:
      - Mission
      operationId: transitionMission
      summary: Executa transição válida da state machine
      parameters:
      - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - to
              properties:
                to:
                  $ref: '#/components/schemas/MissionState'
      responses:
        '200':
          description: Estado alterado
        '409':
          description: Conflito de estado/revisão; reler antes de tentar gravação.
        '403':
          description: Origem, tenant ou papel não autorizados; nunca expõe detalhes internos.
        '413':
          description: Corpo acima de 64 KiB.
        '500':
          description: Falha interna sem vazamento de mensagens do provedor.
...
  /solutions:
    get:
      tags:
      - Solutions
      operationId: listQualifiedSolutions
      summary: 'Calcula providers elegíveis por trusted server cross-tenant service. Sem policy/threshold/mapping
        publicado, retorna estado fail-closed.

        '
      parameters:
      - name: painId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Estado do matching e soluções sanitizadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualifiedSolutionSearch'
        '404':
          description: Feature flag desativada ou pain inexistente
  /solutions/contact:
    post:
      tags:
      - Solutions
      operationId: requestQualifiedSolutionContact
      summary: 'Revalida provider e exige consentimento QUALIFIED_MATCHING efetivo antes de criar oportunidade de
        contato.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - painId
              - providerCapabilityId
              properties:
                painId:
                  type: string
                  format: uuid
                providerCapabilityId:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Solicitação registrada
        '403':
          description: Origem, tenant ou papel não autorizados; nunca expõe detalhes internos.
        '409':
          description: Conflito de estado/revisão; reler antes de tentar gravação.
        '404':
          description: Feature flag desativada
        '413':
          description: Corpo acima de 64 KiB.
        '500':
          description: Falha interna sem vazamento de mensagens do provedor.
...
components:
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Entrada inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Sessão inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
...
    BusinessFactInput:
      type: object
      required:
      - companyId
      - factKey
      - value
      properties:
        companyId:
          type: string
          format: uuid
        factKey:
          type: string
          pattern: ^[a-z0-9_.-]+$
        value: {}
        source:
          type: string
          enum:
          - declared
          default: declared
        sourceRef:
          type: string
        confidence:
          type: 'null'
          description: Confiança não pode ser autodeclarada.
        sensitivity:
          type: string
          enum:
          - public
          - internal
          - personal
          - financial
          - restricted
        purposeCodes:
          type: array
          items:
            type: string
          minItems: 1
      additionalProperties: false
...
    MissionState:
      type: string
      enum:
      - SUGGESTED
      - ACCEPTED
      - IN_PROGRESS
      - EVIDENCE_PENDING
      - COMPLETED
      - OUTCOME_PENDING
      - OUTCOME_RECORDED
      - PAUSED
      - BLOCKED
      - CANCELLED
      - EXPIRED
...
    QualifiedSolution:
      type: object
      required:
      - providerCapabilityId
      - providerCompanyId
      - providerName
      - capabilityId
      - capabilityCode
      - capabilityTitle
      - qualificationScore
      - capacityStatus
      - rankScore
      - explanation
      properties:
        providerCapabilityId:
          type: string
          format: uuid
        providerCompanyId:
          type: string
          format: uuid
        providerName:
          type: string
        providerSector:
          anyOf:
          - type: string
          - type: 'null'
        capabilityId:
          type: string
          format: uuid
        capabilityCode:
          type: string
        capabilityTitle:
          type: string
        qualificationScore:
          type: number
          minimum: 0
          maximum: 100
        capacityStatus:
          type: string
        rankScore:
          type: number
          minimum: 0
          maximum: 1
        explanation:
          type: object
    QualifiedSolutionSearch:
      oneOf:
      - type: object
        required:
        - status
        - painId
        - painCode
        - solutions
        properties:
          status:
            type: string
            enum:
            - POLICY_PENDING
            - NO_CAPABILITY_MAPPING
            - NO_ELIGIBLE_PROVIDERS
          painId:
            type: string
            format: uuid
          painCode:
            type: string
          solutions:
            type: array
            maxItems: 0
      - type: object
        required:
        - status
        - painId
        - painCode
        - solutions
        properties:
          status:
            const: READY
          painId:
            type: string
            format: uuid
          painCode:
            type: string
          solutions:
            type: array
            items:
              $ref: '#/components/schemas/QualifiedSolution'